isoler updateAWSStorage
